refactor: separates common variants
