+++ src/components/Input/BaseInput/index.tsx
@@ -1,26 +1,7 @@
+import InputVariants from '@styles/inputVariants';
 import cn from '@utils/cn';
-import { cva, VariantProps } from 'class-variance-authority';
+import { VariantProps } from 'class-variance-authority';
 import { forwardRef, InputHTMLAttributes } from 'react';
-
-const InputVariants = cva(
-  'h-12 w-full rounded-xl bg-slate-50 font-normal py-3 px-6 placeholder:text-slate-400 border-[1px] outline-none border-transparent ' +
-    'hover:border-blue-300 focus:border-blue-500',
-  {
-    variants: {
-      size: {
-        lg: 'h-12 text-base',
-        sm: 'h-[44px] text-sm',
-      },
-      isInvalid: {
-        true: '!border-red-700',
-        false: '',
-      },
-    },
-    defaultVariants: {
-      isInvalid: false,
-    },
-  },
-);
 
 interface InputProps
   extends Omit<InputHTMLAttributes<HTMLInputElement>, 'size'>,
