+++ src/components/Input/PasswordInput/index.tsx
@@ -1,27 +1,8 @@
 import { VisibilityOffIcon, VisibilityOnIcon } from '@assets';
+import InputVariants from '@styles/inputVariants';
 import cn from '@utils/cn';
-import { cva, VariantProps } from 'class-variance-authority';
+import { VariantProps } from 'class-variance-authority';
 import { forwardRef, InputHTMLAttributes, useState } from 'react';
-
-const InputVariants = cva(
-  'h-12 w-full rounded-xl bg-slate-50 font-normal py-3 pl-6 pr-14 placeholder:text-slate-400 border-[1px] outline-none border-transparent ' +
-    'hover:border-blue-300 focus:border-blue-500',
-  {
-    variants: {
-      size: {
-        lg: 'h-12 text-base',
-        sm: 'h-[44px] text-sm',
-      },
-      isInvalid: {
-        true: '!border-red-700',
-        false: '',
-      },
-    },
-    defaultVariants: {
-      isInvalid: false,
-    },
-  },
-);
 
 interface InputProps
   extends Omit<InputHTMLAttributes<HTMLInputElement>, 'size'>,
@@ -33,23 +14,23 @@ interface InputProps
 // forwardRef<ref가 참조하는 요소 타입, 컴포넌트 props의 타입>
 const PasswordInput = forwardRef<HTMLInputElement, InputProps>(
   ({ size, isInvalid, additionalClass, ...props }, ref) => {
-    const [showPswd, setShowPswd] = useState<boolean>(false);
+    const [showPassword, setShowPassword] = useState<boolean>(false);
     return (
       <div className={cn('relative', additionalClass)}>
         <input
           ref={ref}
-          type={showPswd ? 'text' : 'password'}
+          type={showPassword ? 'text' : 'password'}
           className={cn(InputVariants({ size, isInvalid }))}
           {...props}
         />
-        {showPswd ? (
+        {showPassword ? (
           <VisibilityOnIcon
-            onClick={() => setShowPswd(!showPswd)}
+            onClick={() => setShowPassword((prev) => !prev)}
             className="absolute right-6 top-1/2 -translate-y-1/2 transform cursor-pointer"
           />
         ) : (
           <VisibilityOffIcon
-            onClick={() => setShowPswd(!showPswd)}
+            onClick={() => setShowPassword((prev) => !prev)}
             className="absolute right-6 top-1/2 -translate-y-1/2 transform cursor-pointer"
           />
         )}
